backend/requirements.txt
langchain
faiss-cpu
sentence-transformers
transformers
fastapi
uvicorn
